[PATCH] v4l: msp3400 cleanup.
This patch has some cleanups for the msp3400 module: Balance is used directly
now instead of maintaining the state as left/right volume and calculate the
balance from that.  The msp3400 did that only for historical reasons and it
isn't needed any more ...

Credits for that go to Perry Gilfillan.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 40d64017b3QkdU3F_YD56x6QGmQpOw