[PATCH] larger IO bitmaps
The previous discussion was started here:
http://www.uwsg.iu.edu/hypermail/linux/kernel/0211.0/0477.html but in 2.4
times this was kind of problematic.

Now, with the lazy bitmap allocation and per-CPU TSS, this will really not
drain any resources I think.  8K TSS increase and 8K per process *that does
ioperm()* - I think it is not very bad.

The reasons why I need that, are described in the URL above.  Basically this
will allow to use full-screen VESA under dosemu (without LFB though), and this
may be also helpfull for the XFree project and some other projects:

http://www.uwsg.iu.edu/hypermail/linux/kernel/9807.1/1079.html

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 40df09acD0YLwZXJwyGltO_GSgIuXg