[PATCH] fix GFP zone modifier interators
For each node there are a defined list of MAX_NR_ZONES zones.  These are
selected as a result of the __GFP_DMA and __GFP_HIGHMEM zone modifier flags
being passed to the memory allocator as part of the GFP mask.  Each node
has a set of zone lists, node_zonelists, which defines the list and order
of zones to scan for each flag combination.  When initialising these lists
we iterate over modifier combinations 0 ..  MAX_NR_ZONES.  However, this is
only correct when there are at most ZONES_SHIFT flags.  If another flag is
introduced zonelists for it would not be initialised.

This patch introduces GFP_ZONETYPES (based on GFP_ZONEMASK) as a bound for
the number of modifier combinations.

Signed-off-by: Andy Whitcroft <apw@shadowen.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 40dfbc29spksmw2nAT_sMZRlBnTA1Q