Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/06/28 13:15:54-07:00 kdesler
[NET]: Fix typos in pktgen docs.

2004/06/28 13:14:35-07:00 olh
[DECNET]: Fix signed bug in net/decnet/dn_nsp_in.c:dn_nsp_linkservice()

char can be either signed or unsigned, depending on the target system.

Signed-off-by: Olaf Hering <olh@suse.de>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/06/28 13:10:38-07:00 wli
[AIO]: kiocb->private is too large for kiocb's on-stack.

sizeof(struct kiocb) is dangerously large for a structure commonly
allocated on-stack. This patch converts the 24*sizeof(long) field,
->private, to a void pointer for use by file_operations entrypoints.
A ->dtor() method is added to the kiocb in order to support the release
of dynamically allocated structures referred to by ->private.

The sole in-tree users of ->private are async network read/write,
which are not, in fact, async, and so need not handle preallocated
->private as they would need to if ->ki_retry were ever used. The sole
truly async operations are direct IO pread()/pwrite() which do not
now use ->ki_retry(). All they would need to do in that case is to
check for ->private already being allocated for async kiocbs.

This rips 88B off the stack on 32-bit in the common case.

Signed-off-by: William Lee Irwin III <wli@holomorphy.com>
Signed-off-by: David S. Miller <davem@redhat.com>

BKrev: 40e08ff88M_13UmOziiYwNAdS5-TfQ