[PATCH] sh64: Fix syscall table alignment
If the syscall table is in the text section, the movi that loads its
address into a register will at link time have its immediates resolved as
though it's an SHmedia symbol, i.e.  the LSB will be 1.  This is then
misaligned as a base address for loading the address to jump to.

This could be fixed either by moving the table back to the .data section
(as done here), or by masking off the LSB after loading the table address
into a register.

Signed-off-by: Richard Curnow <richard.curnow@superh.com>
Signed-off-by: Paul Mundt <lethal@linux-sh.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 40e2231eqTlHS8gqy92M8GEfc5Qz4Q