[PATCH] Telephony Driver ISAPNP fix
I recently got an ISA Phonejack card and quickly found it wouldn't work
with a stock 2.6.7 kernel.

So found this bug in the ISAPNP part of the code, throughout the code it
use it's own data structure to reference it's base IO address, in the
ISAPNP code this was not being populated with any value hence it can't talk
to the card and so the driver fails.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 40e59cbceynMAnyWayAN2GfNQhqF7Q