Merge http://linux-mh.bkbits.net/bluetooth-2.6 into nuts.davemloft.net:/disk1/BK/net-2.6
2004/07/04 17:28:19+02:00 marcel
[Bluetooth] Allocate protocol number for AVDTP support

This patch allocates the next free protocol number for the upcoming
AVDTP support.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/04 16:42:38+02:00 marcel
[Bluetooth] Add HID protocol support

This patch adds support for the Bluetooth HID protocol to the
Bluetooth subsystem. Currently only the boot mode is supported.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/04 14:55:47+02:00 marcel
[Bluetooth] Fix kobject oops on firmware loading

Allocate the memory for .kobj.k_name and make use of kobject_init() for
the fake device that is needed for firmware loading.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/04 14:46:07+02:00 marcel
[Bluetooth] Fix deadlock in the 3Com driver

This patch fixes the deadlock problem when bt3c_interrupt() is calling
bt3c_write_wakeup() with info->lock held.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/07/04 12:58:09+02:00 marcel
[Bluetooth] Change inquiry_cache to hci_inquiry_cache

This patch changes the inquiry_* functions to be either local to hci_core.c
or change name from inquiry_cache_lookup to hci_inquiry_cache_lookup to try
and reduce the potential for namespace conflicts. There could be USB or SCSI
or other code that uses an inquiry_cache.

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/06/20 16:18:48+02:00 marcel
[Bluetooth] Fix config change for firmware loading

The switching between firmware loading and running state is done via
different configurations. To get this right the USB toggle and halted
bits must be reset to zero after it.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

2004/06/20 16:12:17+02:00 marcel
[Bluetooth] Fix connection creation error handling

This patch fixes the problem that ACL connections are removed from
the kernel connection hash when a hci_create_connection returns a
command status other than 0x00.

Signed-off-by: Marcel Holtmann <marcel@holtmann.org>

BKrev: 40e9e3ed7UYIuLLQLU7Qky7BAtbFBw