Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2004/07/06 22:02:06-07:00 davem
[IPV4]: Set UDP accept back to sock_no_accept.

Setting it to inet_accept causes UDP accept attempts
to OOPS.  In particular, accept01 from LTP tries this.

Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/06 18:51:12-07:00 davem
[PKT_SCHED]: Kill fix_u32_bug ifdef tests in cls_u32.c

2004/07/06 18:48:46-07:00 davem
[PKT_SCHED]: Kill CONFIG_CLS_U32_PERF2 cruft code.

2004/07/06 16:14:49-07:00 hadi
[PKT_SCHED]: New version of u32 classifier hashing workaround.

This version does not change user visible structures.

Signed-off-by: Jamal Hadi Salim <hadi@znyx.com>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/06 15:37:44-07:00 davem
[PKT_SCHED]: Another missed qdisc_copy_stats() conversion.

2004/07/06 14:35:36-07:00 shemminger
[PKT_SCHED]: Two small netem fixes.

- rtnetlink.h needs including
- optimize loss test so that net_random() call is not done
  when no-loss is indicated

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/06 14:26:44-07:00 davem
[PKT_SCHED]: Fix some missing qdisc_copy_stats() conversions.

2004/07/06 14:05:53-07:00 herbert
[IPCOMP]: Exclude IPCOMP header from props.header_len

This is another patch on the way towards a unified XFRM tunnel
encapsulation function.

This patch changes the value of props.header_len for IPCOMP to
exclude the IPCOMP header.  The reason is that the IPCOMP header
is added only if the packet is compressible.  That is, if the
size of the compressed payload plus the size of the IPCOMP header
is less than that of the original payload.

This means that the IPCOMP encapsulation does not impose any
overhead at all as far as the MTU is concerned.  The current
code incorrectly reduces the MTU by the size of the IPCOMP
header.

As a side-effect, this means that we don't have to move the
IP header around when IPCOMP is used.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/06 13:57:40-07:00 shemminger
[PKT_SCHED]: Update to network emulation QOS scheduler.

This patch updates the network emulation packet scheduler.
	* name changed from delay to netem since it does more than just delay
	* Catalin's merged code to do packet reordering
	* uses a socket queue's directly rather than layering on qdisc(fifo)
	  because this is used in performance tests.
	* adds placeholder in API for future enhancements (rate and duplicate).

Signed-off-by: Stephen Hemminger <shemminger@osdl.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/06 13:50:20-07:00 sebek64
[NETFILTER]: ip6t_LOG and packets with hop-by-hop options.

Packet with IPPROTO_HOPOPTS extended header isn't logged properly by
ip6t_LOG.c. It only prints PROTO=0 and nothing more, because
IPPROTO_HOPOPTS=0 and in this file 0 is used to indicate last header.
This patch fix it by using IPPROTO_NONE to indicate last header.

Signed-off-by: Marcel Sebek <sebek64@post.cz>
Signed-off-by: Harald Welte <laforge@netfilter.org>
Signed-off-by: David S. Miller <davem@redhat.com>

2004/07/05 21:08:20-07:00 davem
[NETFILTER]: Fix initializer of ip6table.c:initial_table

Just because gcc accepts [index] bar as
[index] = bar is no reason to make use of
that gross construct.

Also fully C99ify

2004/07/05 20:39:01-07:00 davem
[PKT_SCHED]: Do not embed spinlock in tc_stats structure.

This makes it not get sized/copied around to/from
userspace correctly.  The real crux of the problem
comes from the rtnetlink attribute copying line which
read:
	RTA_PUT(skb, TCA_STATS, (char*)&st->lock - (char*)st, st);
which is not necessarily sizeof(struct tc_stats) due
to alignment issues.

BKrev: 40eb9906QuvcRuu0QHZky1CGYS6-QA