Back out smc91x late collision "performance" hack.
Apparantly, the reason for this change is that it gives "much less
NFS stalls with a SMC91C96 on a 10/100MB switch."

Alan Cox's comments on this is:
"spec says 16 collision retries max _ever_.  I bet his 'broken
switch' is a case of half v full duplex" and a recommendation to
remove it.

Also, since the change does not limit the number of times we may
retry to send the packet, I think this code opens a potential DoS
attack against network which has a machine connected running this
code.

BKrev: 40ed544dF0ul0Vjm6XWvDYzanZ73aA